libexec/bats: Ensure BATS_READLINK set only once
After running bats with `bash -x bats`, I realized that the
`BATS_READLINK` search was happening with every call to
`resolve_link()`. This ensures it's set only once, which shaves O(0.01s)
off the time of the test suite. Not a huge savings, but the resulting
code is just as clear and works as it's intended.